refactor(pipeline): articles-repo extraction, queue-handler abstraction, discovery via runJsonWithFallback, digest-today lib lane fix
PR2 of the 2026-05-03 ultrareview remediation. Behaviour-preserving
extractions and one ADR; CF-006 phase decomposition deferred to a
PR2b follow-up to keep this PR's blast radius tractable.

CF-002 — discoverTag now routes through runJsonWithFallback so LLM-
assisted source discovery shares the primary→fallback retry, cost
accounting, and waste tracking that the chunk + finalize consumers
already use. The narrow function (narrowDiscoveryPayload) handles
both the string-envelope and already-parsed-object response shapes
that different Workers AI models emit.

CF-004 — loadExistingCanonicalUrls and loadExistingCanonicalToIdMap
move into new src/lib/articles-repo.ts. The Set version is now a
thin wrapper over the Map version (drops one duplicated batched-IN
function and the duplicate EXISTING_URL_BATCH constant).

CF-007 — handleBatch in src/lib/queue-handler.ts consolidates the
identical try/ack/retry/log shape across all three queue consumers
(coordinator, chunk, finalize). MAX_QUEUE_ATTEMPTS sourced from a
single shared constant. Finalize's intentional asymmetry per
REQ-PIPE-008 AC 8 (no finishRun on terminal retry) is expressed as
"no onTerminalFailure passed".

CF-024 — loadTodayPayload + computeNextScrapeAt + the wire types move
from src/pages/api/digest/today.ts into new src/lib/digest-today.ts.
The page (src/pages/digest.astro) was previously importing from a
sibling API route module; both now consume the lib. The route file
keeps re-exports of the wire types for backwards compat with any
test or external consumer that imported them by path.

CF-026 — AD10 (documentation/decisions/README.md) pins the canonical
once-per-run idempotency-gate pattern and explains why no generic
acquireOnceLock helper is introduced (the finalize consumer's gate
fuses with the per-attempt stats fold inside one atomic UPDATE that
a bare lock helper would force-split). The keyed scrape_run_locks
table refactor is the explicit trigger for revisiting when a third
gate would land.

Deferred to PR2b: CF-006 phase decomposition (runCoordinator,
processOneChunk, processOneFinalize), CF-027 broader articles-repo
extraction (insertArticles, appendAlternativeSources,
recordChunkCompletion).
